// src/app/common/models/detail.model.ts
export class Detail {
    number: string;
    first_name: string;
    age: string;
    profile_created_for : string;
    mother_tongue: string;
    weight: string;
    height: String;
    bodyType: string;
    complexion: string;
    smoking_habit: string;
    drinking_habit : string;
    physical_status: string;
    marital_status: string;
    children: String;
    education: string;
    occupation_type: string;
    working_field: string;
    designation : string;
    business: string;
    income: string;
    family_value: String;
    family_type: string;
    father_status: string;
    mother_status: string;
    brother : string;
    brother_married: string;
    sister: string;
    sister_married: string;
    religion: string;
    gothram: string;
    star : string;
    raasi: string;
    caste: string;
    dosham: string;
    birth_hour: string;
    birth_mins: string;
    birth_am_pm: string;
    birth_country: string;
    birth_state: string;
    birth_district: string;
    eating_habit: string;
    photo: string;
}